Revert "Fix interactive plots using `np.arange` (#7924)"
This reverts commit 872ab34c1f99f6fdb98093eeefb1d4c2d069f748.

+++ examples/00-load/create_poly.py
@@ -36,7 +36,7 @@ surf = pv.PolyData(vertices, faces)
 
 # plot each face with a different color
 surf.plot(
-    scalars=np.arange(3, dtype=float),
+    scalars=np.arange(3),
     cpos=[-1, 1, 0.5],
     show_scalar_bar=False,
     show_edges=True,
